Refactored RNN variants to inherit from the same base class
